add some code to a new file scopes/mcp/cli-mcp-server/setup-utils.ts
